# packages/api-client/openapi/reboxe.yaml
openapi: 3.1.0
info:
  title: REBOXE API
  version: 1.4.0
servers:
  - url: http://127.0.0.1:3001
paths:
...
  /v1/kyc/start:
    post:
      operationId: startKyc
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [shopId]
              properties:
                shopId: { type: string }
      responses:
        "201":
          description: Idempotently starts KYC for an owned shop
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KycStartResult" }
  /v1/kyc/document/front:
    post:
      operationId: submitKycDocumentFront
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/KycImage" }
      responses:
        "201":
          description: Normalized front-side OCR and document validation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KycStatus" }
  /v1/kyc/document/back:
    post:
      operationId: submitKycDocumentBack
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/KycImage" }
      responses:
        "201":
          description: Normalized back-side OCR and document validation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KycStatus" }
  /v1/kyc/selfie:
    post:
      operationId: submitKycSelfie
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/KycImage" }
      responses:
        "201":
          description: Normalized face-match and liveness result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KycStatus" }
  /v1/kyc/tax:
    post:
      operationId: submitKycTax
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [kycId, taxCode]
              properties:
                kycId: { type: string }
                taxCode: { type: string, pattern: '^\d{10}(-\d{3})?$' }
      responses:
        "201":
          description: Tax verification recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KycStatus" }
  /v1/kyc/bank:
    post:
      operationId: submitKycBank
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [kycId, bankCode, accountNumber]
              properties:
                kycId: { type: string }
                bankCode: { type: string, minLength: 2, maxLength: 30 }
                accountNumber: { type: string, pattern: '^\d{6,30}$' }
      responses:
        "201":
          description: Bank account verified and authoritative holder name compared
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KycStatus" }
  /v1/kyc/{kycId}/status:
    get:
      operationId: getKycStatus
      security: [{ bearerAuth: [] }]
      parameters:
        - name: kycId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current normalized KYC status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KycStatus" }
...
components:
  requestBodies:
    KycImage:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [kycId, objectKey]
            properties:
              kycId: { type: string }
              objectKey: { type: string, minLength: 1, maxLength: 500 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    KycStartResult:
      allOf:
        - $ref: "#/components/schemas/KycStatus"
        - type: object
          required: [id]
          properties:
            id: { type: string }
    KycReview:
      type: object
      additionalProperties: false
      required: [reason, reviewedAt]
      properties:
        reason: {type: string}
        reviewedAt: {type: string, format: date-time}
    KycIdentity:
      type: object
      additionalProperties: false
      required: [citizenId, fullName, dateOfBirth, gender, address, issuedAt]
      properties:
        citizenId: {type: [string, 'null'], description: Masked identifier; only the last four digits are shown.}
        fullName: {type: [string, 'null']}
        dateOfBirth: {type: [string, 'null']}
        gender: {type: [string, 'null']}
        address: {type: [string, 'null']}
        issuedAt: {type: [string, 'null']}
    KycVerification:
      type: object
      additionalProperties: false
      required: [documentValid, faceMatched, faceScore, livenessPassed, livenessScore]
      properties:
        documentValid: {type: [boolean, 'null']}
        faceMatched: {type: [boolean, 'null']}
        faceScore: {type: [number, 'null']}
        livenessPassed: {type: [boolean, 'null']}
        livenessScore: {type: [number, 'null']}
...
    KycStatus:
      type: object
      additionalProperties: false
      required: [success, kycStatus, review, identity, verification]
      properties:
        success: {type: boolean}
        kycStatus: {type: string, enum: [PENDING, PROCESSING, VERIFIED, REJECTED, MANUAL_REVIEW]}
        review:
          anyOf:
          - {$ref: '#/components/schemas/KycReview'}
          - {type: 'null'}
        identity: {$ref: '#/components/schemas/KycIdentity'}
        verification: {$ref: '#/components/schemas/KycVerification'}
...